Fix for bad zip performance on slaves (#64) The issue is that the zip operation is happening on the master node but the files themselves (zipped files, and the zip file) are on the slave.  This causes poor performance. This fix moves the whole zip operation onto the slave node.
